requirements.txt
# Python 3.9
aiogram==3.0.0
black==23.7.0
PyYAML==6.0.1
openai==0.28.0